Update GraphiQL to version 3.2.0
