+++ assets/explorer/package.json
@@ -3,30 +3,40 @@
   "scripts": {
     "clean": "rimraf dist",
     "build": "npm run clean && NODE_ENV=production webpack && echo \"gzipped, the global build is `gzip -c dist/bundle.min.js | wc -c` bytes\"",
-    "lint": "eslint src"
+    "lint": "eslint src",
+    "dev": "webpack --mode development --watch"
   },
   "dependencies": {
-    "graphiql": "^1.4.2",
-    "graphiql-explorer": "^0.6.3",
-    "graphql": "^15.5.1",
-    "react": "^16.3",
-    "react-dom": "^16.3"
+    "@codemirror/language": "6.0.0",
+    "@graphiql/plugin-explorer": "^2.0.0",
+    "@graphiql/react": "^0.21.0",
+    "graphiql": "^3.2.0",
+    "graphql": "^15.8.0",
+    "graphql-ws": "^5.16.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-is": "^16.13.1"
   },
   "devDependencies": {
-    "babel": "^6.23.0",
-    "babel-core": "^6.26.0",
+    "@babel/cli": "^7.24.1",
+    "@babel/core": "^7.24.4",
+    "@babel/preset-env": "^7.24.4",
+    "@babel/preset-react": "^7.24.1",
+    "@babel/preset-stage-0": "^7.8.3",
+    "@babel/register": "^7.23.7",
     "babel-eslint": "^8.0.2",
-    "babel-loader": "^7.1.2",
+    "babel-loader": "^9.1.3",
     "babel-plugin-syntax-dynamic-import": "^6.18.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
     "copy-webpack-plugin": "^6.1.2",
     "eslint": "^8.9.0",
-    "eslint-config-airbnb": "^18.1.0",
-    "eslint-plugin-import": "^2.20.2",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsx-a11y": "^6.8.0",
+    "eslint-plugin-react": "^7.34.1",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "rimraf": "^2.6.2",
-    "webpack": "^5.40.0",
-    "webpack-cli": "^4.9.2"
+    "webpack": "^5.91.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^5.0.4"
   }
 }
