+++ assets/explorer/src/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import Drupal from 'drupal';
 import jQuery from 'jquery';
 import { buildClientSchema } from 'graphql';
@@ -29,12 +29,15 @@ Drupal.behaviors.graphQLRenderExplorer = {
       },
     }).then(response => response.json());
 
-    // Render <Explorer /> into the container.
-    ReactDOM.render(React.createElement(Explorer, {
-      fetcher: graphQLFetcher,
-      schema: graphQLSchema,
-      query: settings.graphqlQuery || undefined,
-      variables: settings.graphqlVariables || undefined,
-    }), container);
+    const root = createRoot(container);
+
+    root.render(
+      <Explorer
+        fetcher={graphQLFetcher}
+        schema={graphQLSchema}
+        query={settings.graphqlQuery || undefined}
+        variables={settings.graphqlVariables || undefined}
+      />
+    )
   },
 };
